fix(entity-storage): extract translatable fixtures to own files for PSR-4 (#1457)
Root cause: TranslatableArticleFixture and SqlColumnTranslatableArticleFixture
were defined inside the bottom of SqlBlobTranslatableTest.php and
SqlColumnTranslatableTest.php respectively. The classes live in namespace
`Waaseyaa\EntityStorage\Tests\Backend\`, which composer's PSR-4 autoloader
maps to `packages/entity-storage/tests/Backend/<ShortName>.php`. Because the
fixtures had no matching files, the autoloader could not resolve them on
first reference.

Other tests (TranslationWriteSemanticsTest, EntityRepositoryFindTranslationsTest)
referenced these fixtures via `::class`. When PHPUnit happened to load the
parent test file first (alphabetical / Unit-suite-then-Integration), the
fixture was defined as a side effect and worked. When the Integration suite
ran without that side effect — exactly the CI configuration, which runs
`--testsuite Unit` and `--testsuite Integration` as separate phpunit
invocations — `is_subclass_of($class, 'TranslatableInterface')` returned
false for an unloaded class, and `EntityType` constructor threw
InvalidEntityTypeException.

Fix: move each fixture to its own PSR-4-compliant file:
- packages/entity-storage/tests/Backend/TranslatableArticleFixture.php
- packages/entity-storage/tests/Backend/SqlColumnTranslatableArticleFixture.php

Also drop now-unused `ContentEntityType` / `ContentEntityKeys` imports from
the two test files (they only annotated the moved classes). Marked both
fixtures `final` to match the framework convention for concrete test
fixtures.

Verification:
- Full suite: 8264 tests pass (no regression vs prior 8254 baseline).
- Unit suite alone: 7157/7157 (matches CI step 1).
- Integration suite alone: 1097/1097 (matches CI step 2).
- check-composer-policy, check-package-layers, cs-check (two passes),
  phpstan level-5 all green.

Closes #1457.
