+++ packages/entity-storage/tests/Backend/SqlColumnTranslatableTest.php
@@ -9,8 +9,6 @@ use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Waaseyaa\Database\DBALDatabase;
-use Waaseyaa\Entity\Attribute\ContentEntityKeys;
-use Waaseyaa\Entity\Attribute\ContentEntityType;
 use Waaseyaa\Entity\ContentEntityBase;
 use Waaseyaa\Entity\EntityType;
 use Waaseyaa\Entity\EntityTypeManager;
@@ -481,29 +479,5 @@ final class SqlColumnTranslatableTest extends TestCase
             $rows[] = (array) $row;
         }
         return $rows;
-    }
-}
-
-/**
- * @internal Test fixture for {@see SqlColumnTranslatableTest}.
- */
-#[ContentEntityType(id: 'sqlcol_article')]
-#[ContentEntityKeys(
-    id: 'id',
-    uuid: 'uuid',
-    bundle: 'bundle',
-    label: 'label',
-    langcode: 'langcode',
-    default_langcode: 'default_langcode',
-)]
-class SqlColumnTranslatableArticleFixture extends ContentEntityBase
-{
-    public function __construct(
-        array $values = [],
-        string $entityTypeId = '',
-        array $entityKeys = [],
-        array $fieldDefinitions = [],
-    ) {
-        parent::__construct($values, $entityTypeId, $entityKeys, $fieldDefinitions);
     }
 }
